+++ python/lsst/ts/dimm/dimm_csc.py
@@ -28,7 +28,7 @@ from lsst.ts.dimm import controllers
 
 from . import __version__
 from .config_schema import CONFIG_SCHEMA
-from .controllers.base_dimm import DIMMStatus
+from .controllers.base_dimm import AutomationMode, DIMMStatus
 from .utils.conversion import (
     convert_dimm_measurement_data,
     convert_to_float,
@@ -237,11 +237,6 @@ class DIMMCSC(salobj.ConfigurableCsc):
         self.telemetry_loop_running = False
         self.seeing_loop_running = False
 
-        try:
-            await self.controller.stop()
-        except Exception:
-            self.log.exception("Error in begin_disable. Continuing...")
-
         await super().begin_disable(id_data)
 
     async def end_disable(self, id_data):
@@ -266,6 +261,11 @@ class DIMMCSC(salobj.ConfigurableCsc):
             await self.wait_loop(self.seeing_loop_task)
         except Exception:
             self.log.exception("Error trying to stop the seeing loop. Continuing.")
+
+        try:
+            await self.controller.stop()
+        except Exception:
+            self.log.exception("Error in controller stop. Continuing...")
 
         await super().end_disable(id_data)
 
@@ -340,9 +340,14 @@ class DIMMCSC(salobj.ConfigurableCsc):
                     )
                     break
 
+                # TODO: DM-48873 Remove the `hasattr` condition when cycle 39
+                # XML is no longer in service.
+                if hasattr(self, "tel_ameba"):
+                    self.log.debug("Collecting AMEBA state.")
+                    ameba_topic = await self.controller.get_ameba()
+                    await self.tel_ameba.set_write(**ameba_topic)
+
                 await asyncio.sleep(self.heartbeat_interval)
-        except asyncio.CancelledError:
-            pass
         except Exception:
             self.log.exception("Error in telemetry loop.")
             await self.fault(
@@ -487,7 +492,8 @@ class DIMMCSC(salobj.ConfigurableCsc):
             Contains the data as defined in the SAL XML file.
         """
         self.assert_enabled()
-        raise salobj.ExpectedError("Not implemented yet.")
+        mode = AutomationMode(data.mode)
+        await self.controller.set_automation_mode(mode)
 
     async def do_changeDwellRate(self, data):
         """Change how long to sample for in seconds.